Make the dock more accurate
